papers/Learn_to_Abstain/requirements.txt
h5py==3.9.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
numba==0.58.1
numpy==1.26.0
opencv-python==4.8.1.78
pandas==2.1.1
scikit-learn==1.2.2
scipy==1.11.4
statsmodels==0.13.5
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
termcolor==2.3.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.1
pyyaml==6.0.1